+++ commands/help.js
@@ -1,0 +1,7 @@
+module.exports = {
+	name: 'help',
+	description: 'Help!',
+	execute(message, args) {
+		message.channel.send('help.');
+	},
+};
